Separa o componente "barOder" que a calculadora para os adicionais da pagina de detalhes do item onde é possivel montar o hamburguer com adicionais e inicia a correção do erro ao clicar em detalhes de adicionais da pagina menu
