+++ frontend-menu/src/style/itemDetails.css
@@ -144,7 +144,7 @@
   flex-direction: column;
   background-color: black;
   padding: 0;
-  padding-top: 15px;
+  padding-top: 20px;
 }
 
 .div_tags_buy {
